% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thresholddata.R
\name{extract_thresholddata}
\alias{extract_thresholddata}
\title{A function which extracts tresholdData that was measured with USZ_STT}
\usage{
extract_thresholddata(list_datafiles)
}
\arguments{
\item{list_datafiles}{A list with xls(x) files you would like to read in and extract}
}
\description{
This function allows to read in thresholData (xlsx format) from multiple participants
and converts it into a single dataframe
}
\examples{
thr_filelist <- list.files(path = paste0(inputFolder, "/Audiometrie/"), pattern = "Thr .*\\\\.xls$", full.names = TRUE, recursive = TRUE)

extract_thresholddata(thr_filelist)
}
\keyword{data}
\keyword{extraction}
